requirements.txt
# Real Estate Walkthrough Bot — Python dependencies
#
# System dependency (NOT pip-installable — install separately):
#   ffmpeg  →  OGG → 192kbps MP3 conversion before Whisper transcription
#   Local:   https://ffmpeg.org/download.html
#   Railway: handled automatically via nixpacks.toml

python-telegram-bot>=21.0
openai>=1.0.0
google-genai>=0.8.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydub>=0.25.1
